restructure json_like
