+++ src/core/json/graphql.rs
@@ -1,0 +1,121 @@
+use std::collections::HashMap;
+
+use async_graphql::Name;
+use async_graphql_value::ConstValue;
+use indexmap::IndexMap;
+
+use super::*;
+
+// ConstValue
+impl<V: JsonLike + Clone> JsonObjectLike for IndexMap<async_graphql_value::Name, V> {
+    type Value<'a> = V where V: 'a;
+    fn get_key<'a>(&'a self, key: &str) -> Option<&Self::Value<'a>> {
+        self.get(&async_graphql_value::Name::new(key))
+    }
+}
+
+impl JsonLike for ConstValue {
+    type JsonObject<'a> = IndexMap<Name, ConstValue>;
+    type JsonArray<'a> = Vec<ConstValue>;
+
+    fn as_array_ok<'a>(&'a self) -> Result<&Self::JsonArray<'a>, &str> {
+        match self {
+            ConstValue::List(seq) => Ok(seq),
+            _ => Err("array"),
+        }
+    }
+
+    fn as_str_ok(&self) -> Result<&str, &str> {
+        match self {
+            ConstValue::String(s) => Ok(s),
+            _ => Err("str"),
+        }
+    }
+
+    fn as_i64_ok(&self) -> Result<i64, &str> {
+        match self {
+            ConstValue::Number(n) => n.as_i64().ok_or("expected i64"),
+            _ => Err("i64"),
+        }
+    }
+
+    fn as_u64_ok(&self) -> Result<u64, &str> {
+        match self {
+            ConstValue::Number(n) => n.as_u64().ok_or("expected u64"),
+            _ => Err("u64"),
+        }
+    }
+
+    fn as_f64_ok(&self) -> Result<f64, &str> {
+        match self {
+            ConstValue::Number(n) => n.as_f64().ok_or("expected f64"),
+            _ => Err("f64"),
+        }
+    }
+
+    fn as_bool_ok(&self) -> Result<bool, &str> {
+        match self {
+            ConstValue::Boolean(b) => Ok(*b),
+            _ => Err("bool"),
+        }
+    }
+
+    fn as_null_ok(&self) -> Result<(), &str> {
+        match self {
+            ConstValue::Null => Ok(()),
+            _ => Err("null"),
+        }
+    }
+
+    fn as_option_ok(&self) -> Result<Option<&Self>, &str> {
+        match self {
+            ConstValue::Null => Ok(None),
+            _ => Ok(Some(self)),
+        }
+    }
+
+    fn get_path<T: AsRef<str>>(&self, path: &[T]) -> Option<&Self> {
+        let mut val = self;
+        for token in path {
+            val = match val {
+                ConstValue::List(seq) => {
+                    let index = token.as_ref().parse::<usize>().ok()?;
+                    seq.get(index)?
+                }
+                ConstValue::Object(map) => map.get(token.as_ref())?,
+                _ => return None,
+            };
+        }
+        Some(val)
+    }
+
+    fn get_key(&self, path: &str) -> Option<&Self> {
+        match self {
+            ConstValue::Object(map) => map.get(&async_graphql::Name::new(path)),
+            _ => None,
+        }
+    }
+
+    fn group_by<'a>(&'a self, path: &'a [String]) -> HashMap<String, Vec<&'a Self>> {
+        let src = gather_path_matches(self, path, vec![]);
+        group_by_key(src)
+    }
+
+    fn null() -> Self {
+        Default::default()
+    }
+
+    fn as_object_ok<'a>(&'a self) -> Result<&Self::JsonObject<'a>, &str> {
+        match self {
+            ConstValue::Object(map) => Ok(map),
+            _ => Err("expected object"),
+        }
+    }
+}
+
+impl JsonArrayLike for Vec<ConstValue> {
+    type Value<'a> = ConstValue;
+    fn as_vec<'a>(&'a self) -> &'a Vec<&Self::Value<'a>> {
+        todo!()
+    }
+}
