+++ src/core/json/mod.rs
@@ -1,7 +1,61 @@
+mod borrow;
+mod graphql;
 mod json_like;
-mod json_object_like;
 mod json_schema;
+mod serde;
 
+use std::collections::HashMap;
+
+pub use borrow::*;
+pub use graphql::*;
 pub use json_like::*;
-pub use json_object_like::*;
 pub use json_schema::*;
+pub use serde::*;
+
+// Highly micro-optimized and benchmarked version of get_path_all
+// Any further changes should be verified with benchmarks
+fn gather_path_matches<'a, J: JsonLike>(
+    root: &'a J,
+    path: &'a [String],
+    mut vector: Vec<(&'a J, &'a J)>,
+) -> Vec<(&'a J, &'a J)>
+where
+    J::JsonObject<'a>: JsonObjectLike<Value<'a> = J>,
+    J::JsonArray<'a>: JsonArrayLike<Value<'a> = J>,
+{
+    if let Ok(root) = root.as_array_ok() {
+        for value in root.as_vec().iter() {
+            vector = gather_path_matches(value, path, vector);
+        }
+    } else if let Some((key, tail)) = path.split_first() {
+        if let Some(value) = root.get_key(key) {
+            if tail.is_empty() {
+                vector.push((value, root));
+            } else {
+                vector = gather_path_matches(value, tail, vector);
+            }
+        }
+    }
+
+    vector
+}
+
+fn group_by_key<'a, J: JsonLike>(src: Vec<(&'a J, &'a J)>) -> HashMap<String, Vec<&'a J>> {
+    let mut map: HashMap<String, Vec<&'a J>> = HashMap::new();
+    for (key, value) in src {
+        // Need to handle number and string keys
+        let key_str = key
+            .as_str_ok()
+            .map(|a| a.to_string())
+            .or_else(|_| key.as_f64_ok().map(|a| a.to_string()));
+
+        if let Ok(key) = key_str {
+            if let Some(values) = map.get_mut(&key) {
+                values.push(value);
+            } else {
+                map.insert(key, vec![value]);
+            }
+        }
+    }
+    map
+}
